Allow the user to specify a location to save uploaded files permanently

+++ threadfix-main/src/main/java/com/denimgroup/threadfix/webapp/controller/UploadScanController.java
@@ -26,11 +26,13 @@ package com.denimgroup.threadfix.webapp.controller;
 import com.denimgroup.threadfix.DiskUtils;
 import com.denimgroup.threadfix.data.ScanCheckResultBean;
 import com.denimgroup.threadfix.data.ScanImportStatus;
+import com.denimgroup.threadfix.data.entities.DefaultConfiguration;
 import com.denimgroup.threadfix.data.entities.Organization;
 import com.denimgroup.threadfix.data.entities.Permission;
 import com.denimgroup.threadfix.data.entities.Scan;
 import com.denimgroup.threadfix.importer.interop.ScanTypeCalculationService;
 import com.denimgroup.threadfix.logging.SanitizedLogger;
+import com.denimgroup.threadfix.service.DefaultConfigService;
 import com.denimgroup.threadfix.service.OrganizationService;
 import com.denimgroup.threadfix.service.ScanMergeService;
 import com.denimgroup.threadfix.service.ScanService;
@@ -60,11 +62,13 @@ public class UploadScanController {
     @Autowired
     private ScanTypeCalculationService scanTypeCalculationService;
     @Autowired
-    private ScanService                scanService;
+    private ScanService                 scanService;
     @Autowired
-    private ScanMergeService           scanMergeService;
+    private ScanMergeService            scanMergeService;
     @Autowired
-    private OrganizationService        organizationService;
+    private OrganizationService         organizationService;
+    @Autowired
+    private DefaultConfigService        defaultConfigService;
 
     @RequestMapping(value = "/organizations/{orgId}/applications/{appId}/upload/remote", method = RequestMethod.POST, produces = "text/plain")
     @JsonView(AllViews.TableRow.class)
@@ -122,16 +126,23 @@ public class UploadScanController {
                 return failure(returnValue.getScanCheckResult().toString());
             }
         } finally { // error recovery code
-            File diskFile = DiskUtils.getScratchFile(fileName);
 
-            if (diskFile.exists()) {
-                LOG.info("After scan upload, file is still present. Attempting to delete.");
-                boolean deletedSuccessfully = diskFile.delete();
+            DefaultConfiguration defaultConfig = defaultConfigService.loadCurrentConfiguration();
+            String fileUploadLocation = defaultConfig.getFileUploadLocation();
 
-                if (deletedSuccessfully) {
-                    LOG.info("Successfully deleted scan file.");
-                } else {
-                    LOG.error("Unable to delete file.");
+            if(fileUploadLocation == null || fileUploadLocation.isEmpty()) {
+
+                File diskFile = DiskUtils.getScratchFile(fileName);
+
+                if (diskFile.exists()) {
+                    LOG.info("After scan upload, file is still present. Attempting to delete.");
+                    boolean deletedSuccessfully = diskFile.delete();
+
+                    if (deletedSuccessfully) {
+                        LOG.info("Successfully deleted scan file.");
+                    } else {
+                        LOG.error("Unable to delete file.");
+                    }
                 }
             }
         }
